requirements.txt
dotenv
requests
openai
pillow
fastapi